+++ regen/effects/bloom-pass.h
@@ -1,7 +1,7 @@
 #ifndef REGEN_BLOOM_PASS_H
 #define REGEN_BLOOM_PASS_H
 
-#include <regen/states/fullscreen-pass.h>
+#include "fullscreen-pass.h"
 #include "regen/textures/fbo-state.h"
 #include "bloom-texture.h"
