+++ regen/effects/geometric-picking.cpp
