+++ regen/scene/horizon-divider.cpp
