+++ regen/scene/horizon-divider.h
